+++ bar/bar.h
@@ -17,12 +17,14 @@ struct bar {
 };
 
 enum bar_location { BAR_TOP, BAR_BOTTOM };
+enum bar_layer { BAR_LAYER_TOP, BAR_LAYER_BOTTOM };
 enum bar_backend { BAR_BACKEND_AUTO, BAR_BACKEND_XCB, BAR_BACKEND_WAYLAND };
 
 struct bar_config {
     enum bar_backend backend;
 
     const char *monitor;
+    enum bar_layer layer;
     enum bar_location location;
     int height;
     int left_spacing, right_spacing;