+++ bar/wayland.c
@@ -988,10 +988,14 @@ setup(struct bar *_bar)
 
     wl_surface_add_listener(backend->surface, &surface_listener, backend);
 
+    enum zwlr_layer_shell_v1_layer layer = bar->layer == BAR_LAYER_BOTTOM
+        ? ZWLR_LAYER_SHELL_V1_LAYER_BOTTOM
+        : ZWLR_LAYER_SHELL_V1_LAYER_TOP;
+
     backend->layer_surface = zwlr_layer_shell_v1_get_layer_surface(
         backend->layer_shell, backend->surface,
         backend->monitor != NULL ? backend->monitor->output : NULL,
-        ZWLR_LAYER_SHELL_V1_LAYER_BOTTOM, "panel");
+        layer, "panel");
 
     if (backend->layer_surface == NULL) {
         LOG_ERR("failed to create layer shell surface");